RichMedia Ad
This feature was designed as a sales component for the Hubbard
Broadcasting websites.  The file,
public.js, is the front-facing javascript file that retrieves the
information via an ajax call and renders the richmedia
ad.  The Ad appears below the main menu.  The add allows for a main
graphic and a video to go in the ad.  Furthermore,
there is the ability to create a parallax ad.  This is a utilizing a
jQuery plugin that layers multiple .pngs on top
of a .jpeg background and allows them to scroll at different speeds /
layer.